django_app/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
drf-spectacular==0.28.0

psycopg2-binary==2.9.7
PyMySQL==1.0.3

python-dotenv==1.0.1
kafka-python==2.0.2
requests==2.31.0
gunicorn==20.1.0
